Big update
Made the camea follow the player only when moving out of 'bounds'.
moving right means the player can be all over the left side w/o moving
the camera, and vice versa.  added in some slow moving clouds to give
the background a bit of pizzaz.  added in a rough AudioManager class
component firstly to play a jump noise. Also, removed the
movingsimplesprites class since it was unneeded and something i'd
creaded earlier.